# expression: implement vectorized evaluation for `builtinYearWeekWithoutModeSig`