<!--
Thank you for contributing to TiDB! Please read TiDB's [CONTRIBUTING](https://github.com/pingcap/community/blob/master/CONTRIBUTING.md) document **BEFORE** filing this PR.
-->

PCP #12103 

### What problem does this PR solve? <!--add issue link with summary if exists-->

vectorize builtinYearWeekWithoutModeSig, for #12103 

### What is changed and how it works?

```
make vectorized-bench VB_FILE=Time VB_FUNC=builtinYearWeekWithoutModeSig
cd ./expression && \
                go test -v -benchmem \
                        -bench=BenchmarkVectorizedBuiltinTimeFunc \
                        -run=BenchmarkVectorizedBuiltinTimeFunc \
                        -args "builtinYearWeekWithoutModeSig"
goos: darwin
goarch: amd64
pkg: github.com/pingcap/tidb/expression
BenchmarkVectorizedBuiltinTimeFuncGenerated-8           1000000000               0.00760 ns/op         0 B/op          0 allocs/op
BenchmarkVectorizedBuiltinTimeFunc/builtinYearWeekWithoutModeSig-VecBuiltinFunc-8                  37816             31018 ns/op               0 B/op          0 allocs/op
BenchmarkVectorizedBuiltinTimeFunc/builtinYearWeekWithoutModeSig-NonVecBuiltinFunc-8               24514             47851 ns/op               0 B/op          0 allocs/op
PASS
ok      github.com/pingcap/tidb/expression      3.253s
```

### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

 - Unit test